feat(images): add document preview content validation
