+++ db/model/Gdoc/gdocUtils.ts
@@ -77,22 +77,3 @@ export const getTitleSupertitleFromHeadingText = (
         afterSeparator ? beforeSeparator : undefined,
     ]
 }
-
-export function recursivelyMapArticleBlock(
-    block: OwidEnrichedArticleBlock,
-    callback: (block: OwidEnrichedArticleBlock) => OwidEnrichedArticleBlock
-): OwidEnrichedArticleBlock {
-    if (block.type === "gray-section") {
-        block.items.map((block) => recursivelyMapArticleBlock(block, callback))
-    }
-    if (
-        block.type === "sticky-left" ||
-        block.type === "sticky-right" ||
-        block.type === "side-by-side"
-    ) {
-        block.left.map((node) => recursivelyMapArticleBlock(node, callback))
-        block.right.map((node) => recursivelyMapArticleBlock(node, callback))
-    }
-
-    return callback(block)
-}
